+++ frontend/src/app/components/customer-all-borrowings/customer-all-borrowings.component.ts
@@ -18,6 +18,7 @@ export class CustomerAllBorrowingsComponent implements OnInit{
 
   borrowings: BorrowingDTO[] = [];
   azonosito: number = 0;
+  nev: string = '';
 
   errorMessage: string = "";
 
@@ -27,7 +28,14 @@ export class CustomerAllBorrowingsComponent implements OnInit{
     this.azonosito = Number(this.activatedroute.snapshot.paramMap.get('azonosito'));
 
     this.borrowingService.getCustomerBorrowings(this.azonosito).subscribe({
-      next: (borrowings) => this.borrowings = borrowings,
+      next: (borrowings) => {
+        this.borrowings = borrowings
+        if (this.borrowings.length > 0) {
+          this.nev = this.borrowings[0].customer.nev;
+        } else {
+          this.nev = 'Ismeretlen ügyfél';
+        }
+      },
       error: (err) => {
         console.error('Failed to load borrowings', err)
         this.errorMessage = err.error.message;