+++ frontend/src/app/components/customer-datasheet/customer-datasheet.component.ts
@@ -102,6 +102,7 @@ export class CustomerDatasheetComponent implements OnInit{
     this.showUpdatePopup = false;
     this.fieldToUpdate = '';
     this.newValue = '';
+    this.errorMessage = '';
   }
 
   updateCustomer() {