+++ frontend/src/app/components/media-datasheet/media-datasheet.component.ts
@@ -70,6 +70,7 @@ export class MediaDatasheetComponent implements OnInit{
       this.showUpdatePopup = false;
       this.fieldToUpdate = '';
       this.newValue = '';
+      this.errorMessage = '';
     }
 
     updateCustomer() {